+++ clang-tools-extra/clang-tidy/hooks/HooksTidyModule.cpp
@@ -25,6 +25,7 @@
 #include "FieldAddBufLenCheck.h"
 #include "FieldBufLenCheck.h"
 #include "FieldDelBufLenCheck.h"
+#include "FloatOnePureCheck.h"
 #include "FloatPureCheck.h"
 #include "GuardCalledCheck.h"
 #include "GuardInForCheck.h"
@@ -76,6 +77,8 @@ public:
         "hooks-field-buf-len");
     CheckFactories.registerCheck<FieldDelBufLenCheck>(
         "hooks-field-del-buf-len");
+    CheckFactories.registerCheck<FloatOnePureCheck>(
+        "hooks-float-one-pure");
     CheckFactories.registerCheck<FloatPureCheck>(
         "hooks-float-pure");
     CheckFactories.registerCheck<GuardCalledCheck>(